Move LocationService to location package.
